training/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
scipy>=1.10.0
pyyaml>=6.0
pyarrow>=10.0.0
